Further development of the trading system. You might be able to sell stuff now. Some fixes and tweaks included as well.

+++ C# Text Adventure/Items/Armor.cs
@@ -3,7 +3,7 @@ public class ArmorItem : Item
 {
     public override string Name => Color.FORE_ORANGE + RawName + Color.RESET;
     public int Defense { get; private init; }
-    internal ArmorItem(string name, string description, double weight, double value, int defense)
+    internal ArmorItem(string name, string description, double weight, int value, int defense)
         : base(name, description, weight, value)
     {
         Defense = defense;
