+++ C# Text Adventure/Items/Healing.cs
@@ -2,7 +2,7 @@
 public class HealingItem : Item
 {
     public int HealAmount { get; private init; }
-    internal HealingItem(string name, string description, double weight, double value, int healAmount)
+    internal HealingItem(string name, string description, double weight, int value, int healAmount)
         : base(name, description, weight, value)
     {
         HealAmount = healAmount;
