+++ C# Text Adventure/Items/Story.cs
@@ -2,7 +2,7 @@
 public class InfoItem : Item
 {
     public string Message { get; private init; }
-    internal InfoItem(string name, string description, double weight, double value, string message)
+    internal InfoItem(string name, string description, double weight, int value, string message)
         : base(name, description, weight, value)
     {
         Message = message;
