+++ C# Text Adventure/Items/Weapon.cs
@@ -3,7 +3,7 @@ public class WeaponItem : Item
 {
     public override string Name => Color.FORE_LIGHT_PURPLE + RawName + Color.RESET;
     public int Damage { get; private init; }
-    internal WeaponItem(string name, string description, double weight, double value, int damage)
+    internal WeaponItem(string name, string description, double weight, int value, int damage)
         : base(name, description, weight, value)
     {
         Damage = damage;
